- Modified StatusInfo to take an int32 value instead of an int16  - Moved the casting for StatusInfo from all the declarations to the packets  - Modified MobDataProvider to give skill data (in the event that linked mobs also inherit skills?), and regardless, it makes the skill routines snappier  - Modified mob skill summons to spread out a little if there are multiples
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@2494 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/MobsPacket.cpp
@@ -135,7 +135,7 @@ void MobsPacket::applyStatus(Mob *mob, int32_t statusmask, const vector<StatusIn
 	packet.add<int32_t>(statusmask);
 
 	for (size_t i = 0; i < info.size(); i++) {
-		packet.add<int16_t>(info[i].val);
+		packet.add<int16_t>(static_cast<int16_t>(info[i].val));
 		if (info[i].skillid >= 0) {
 			packet.add<int32_t>(info[i].skillid);
 		}
